clippy + fmt
